remove alasql refactor saveusersessions
