+++ node/node-notify-server/package.json
@@ -5,17 +5,16 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
+    "dev": "node --inspect server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Daniel Hochleitner",
   "license": "MIT",
   "dependencies": {
-    "alasql": "^0.2.7",
     "encoding": "^0.1.12",
     "fs": "0.0.2",
     "http": "0.0.0",
     "ioredis": "^5.3.2",
-    "redis": "^4.6.10",
     "redisocket": "^4.0.1",
     "socket.io": "^1.7.4",
     "url": "^0.11.0"
